Big reimport for compatibility with TMPro
For some reason, despite having TMPro installed through the PackageManager, I was unable to access it in code. So I had to create a new project where it DID work, then copy everything over to this so things would work.

Also did some reorganizing of files, like with the demo scene.